# nestjs-project/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/videos": {
      "post": {
        "description": "Creates the video as a draft in the caller channel, starts an S3 multipart upload and returns one presigned URL per part so the client uploads directly to storage.",
        "operationId": "VideosController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateVideoDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoDraftResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "413": {
            "description": "Video exceeds the size limit",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Pre-register a video and start its upload",
        "tags": [
          "videos"
        ]
      }
    },
    "/videos/{id}/upload": {
      "get": {
        "description": "Lists the parts already stored and returns fresh presigned URLs for the missing ones, so an interrupted upload can be resumed.",
        "operationId": "VideosController_getUploadState",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadStateResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Video not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Video is not a draft",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get the upload state of a draft",
        "tags": [
          "videos"
        ]
      }
    },
    "/videos/{id}/upload/complete": {
      "post": {
        "description": "Completes the multipart upload, checks the stored size, moves the video to processing and enqueues the processing job. Repeating the call while processing is idempotent.",
        "operationId": "VideosController_completeUpload",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteUploadDto"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoProcessingResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid body, id or upload parts",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Video not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Video is ready or failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "413": {
            "description": "Stored object exceeds the size limit",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Complete the upload and start processing",
        "tags": [
          "videos"
        ]
      }
    },
    "/videos/{id}": {
      "get": {
        "description": "Returns the video status lifecycle (draft, processing, ready, failed), extracted metadata and a presigned thumbnail URL once a thumbnail exists.",
        "operationId": "VideosController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Video not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get a video owned by the caller",
        "tags": [
          "videos"
        ]
      },
      "delete": {
        "description": "Aborts the multipart upload of a draft video and removes it. Only drafts can be deleted in this phase.",
        "operationId": "VideosController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Draft removed"
          },
          "400": {
            "description": "Invalid id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Video not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Video is not a draft",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Abort an unfinished upload",
        "tags": [
          "videos"
        ]
      }
    },
    "/videos/{publicId}/stream": {
      "get": {
        "description": "Redirects (302) to a presigned storage URL; storage serves Range requests with 206 Partial Content, so playback starts without a full download.",
        "operationId": "VideosController_stream",
        "parameters": [
          {
            "name": "publicId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "Redirect to the presigned URL"
          },
          "404": {
            "description": "Unknown video or video not ready",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests"
          }
        },
        "summary": "Stream a ready video",
        "tags": [
          "videos"
        ]
      }
    },
    "/videos/{publicId}/download": {
      "get": {
        "description": "Redirects (302) to a presigned storage URL that forces Content-Disposition: attachment with the original file name.",
        "operationId": "VideosController_download",
        "parameters": [
          {
            "name": "publicId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "Redirect to the presigned URL"
          },
          "404": {
            "description": "Unknown video or video not ready",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "Too many requests"
          }
        },
        "summary": "Download a ready video",
        "tags": [
          "videos"
        ]
      }
    }
  },
  "info": {
    "title": "StreamTube API",
    "description": "API REST do StreamTube",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateVideoDto": {
        "type": "object",
        "properties": {}
      },
      "PresignedPartDto": {
        "type": "object",
        "properties": {
          "part_number": {
            "type": "number",
            "example": 1
          },
          "url": {
            "type": "string",
            "example": "http://localhost:9000/streamtube-videos/..."
          }
        },
        "required": [
          "part_number",
          "url"
        ]
      },
      "UploadInstructionsDto": {
        "type": "object",
        "properties": {
          "part_size": {
            "type": "number",
            "example": 104857600
          },
          "part_count": {
            "type": "number",
            "example": 3
          },
          "parts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PresignedPartDto"
            }
          },
          "expires_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "part_size",
          "part_count",
          "parts",
          "expires_at"
        ]
      },
      "VideoDraftResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "public_id": {
            "type": "string",
            "example": "aZ3kP9xQ2mB"
          },
          "title": {
            "type": "string",
            "example": "My video"
          },
          "status": {
            "type": "string",
            "example": "draft"
          },
          "upload": {
            "$ref": "#/components/schemas/UploadInstructionsDto"
          }
        },
        "required": [
          "id",
          "public_id",
          "title",
          "status",
          "upload"
        ]
      },
      "UploadedPartDto": {
        "type": "object",
        "properties": {
          "part_number": {
            "type": "number",
            "example": 1
          },
          "etag": {
            "type": "string",
            "example": "\"5d41402abc4b2a76b9719d911017c592\""
          },
          "size": {
            "type": "number",
            "example": 104857600
          }
        },
        "required": [
          "part_number",
          "etag",
          "size"
        ]
      },
      "UploadStateResponseDto": {
        "type": "object",
        "properties": {
          "part_size": {
            "type": "number",
            "example": 104857600
          },
          "part_count": {
            "type": "number",
            "example": 3
          },
          "uploaded_parts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UploadedPartDto"
            }
          },
          "parts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PresignedPartDto"
            }
          },
          "expires_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "part_size",
          "part_count",
          "uploaded_parts",
          "parts",
          "expires_at"
        ]
      },
      "CompleteUploadDto": {
        "type": "object",
        "properties": {}
      },
      "VideoProcessingResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string",
            "example": "processing"
          }
        },
        "required": [
          "id",
          "status"
        ]
      },
      "VideoResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "public_id": {
            "type": "string",
            "example": "aZ3kP9xQ2mB"
          },
          "title": {
            "type": "string",
            "example": "My video"
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "processing",
              "ready",
              "failed"
            ]
          },
          "original_filename": {
            "type": "string",
            "example": "clip.mp4"
          },
          "content_type": {
            "type": "string",
            "example": "video/mp4"
          },
          "size_bytes": {
            "type": "number",
            "nullable": true,
            "example": 11534336
          },
          "duration_seconds": {
            "type": "number",
            "nullable": true,
            "example": 3
          },
          "width": {
            "type": "number",
            "nullable": true,
            "example": 1280
          },
          "height": {
            "type": "number",
            "nullable": true,
            "example": 720
          },
          "video_codec": {
            "type": "string",
            "nullable": true,
            "example": "h264"
          },
          "thumbnail_url": {
            "type": "string",
            "nullable": true,
            "description": "Presigned thumbnail URL; null until a thumbnail exists"
          },
          "processing_error": {
            "type": "string",
            "nullable": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "public_id",
          "title",
          "status",
          "original_filename",
          "content_type",
          "size_bytes",
          "duration_seconds",
          "width",
          "height",
          "video_codec",
          "thumbnail_url",
          "processing_error",
          "created_at",
          "updated_at"
        ]
      },
      "ApiErrorEnvelope": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 401
          },
          "error": {
            "type": "string",
            "example": "INVALID_CREDENTIALS"
          },
          "message": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ],
            "example": "Invalid email or password"
          },
          "code": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "statusCode",
          "error",
          "message"
        ]
      }
    }
  }
}